Fix some casting issues on 64 bit archs.

+++ SpecTcl/Display/refreshctl.cc
@@ -376,7 +376,7 @@ static void CheckAbort(XtPointer userd, XtIntervalId *id)
   /*
   ** First determine which pane we're dealing with.
   */
-  int index = (int)userd;
+  long index = (long)userd;
   int column = index  % WINDOW_MAXAXIS;
   int row    = column / WINDOW_MAXAXIS;
